Un-expose test-only public API (Phase 3)
Downgrades from_bom (HourlyForecast, DailyForecast) and into_domain
(OpenMeteoHourlyResponse, OpenMeteoDailyResponse) from pub to
pub(crate): now that Phase 2 removed their only external-crate callers
(the deleted tests/*.rs files), pub was only serving the test crates'
visibility needs, not any real API contract. providers/bom.rs and
providers/open_meteo.rs still call them as in-crate callers.

Makes Fetcher::classify_error/is_error_retryable/parse_retry_after
private and moves their 15 direct tests into a #[cfg(test)] module in
fetcher.rs (wiremock is already a dev-dependency, so tests needing a
real reqwest::Error still get one from a local mock server). Two tests
misfiled in the same integration test file actually exercised
DashboardError::priority()/Description directly, unrelated to fetcher
wiring, so those move to a new errors.rs unit-test module instead.
tests/fetcher_test.rs is trimmed to the 3 genuine end-to-end
retry/backoff wiring tests and its unused mod helpers import is
dropped.

Full suite: 185 passed before and after.
